[Enhance] combine the EvmTx interface.
