+++ src/features/EVM/Transaction/components/Main.tsx
@@ -11,17 +11,11 @@ import Field, { FieldLine } from "src/components/Field";
 import Header from "src/components/Header";
 import { useTransactionMachine } from "src/machines/transaction";
 import { logSendTx } from "src/services/Amplitude";
+import { EvmTransaction } from "src/types";
 import useTransactionDetail from "../hooks/useTransactionDetail";
 import messages from "./messages";
 import TransactionContent from "./TransactionContent";
 import TransactionInfo from "./TransactionInfo";
-
-export interface EvmTransaction {
-  data: string;
-  from: string;
-  to: string;
-  value: string;
-}
 
 const Main = () => {
   const { context, send } = useTransactionMachine();
